// api_v2/src/enums/input-validation-message.enum.ts
export enum InputValidationMessageEnum {
  ALREADY_USED_ESL_OFFSET = 'Already used ESL Offset ID',
  DUPLICATE_EMPLOYEE = ' is already in the schedule!',
  DUPLICATE_REQUEST = "There's already an existing request!",
  DUPLICATE_SCHEDULE_NAME = 'Schedule name already exists!',
  INVALID_DATE = 'Invalid Date',
  INVALID_DAY = 'Invalid Day',
  INVALID_EMAIL = 'Invalid Email',
  INVALID_EMPLOYEE = "Employee doesn't exist",
  INVALID_END_TIME = 'Invalid End Time',
  INVALID_ESL_OFFSET_IDS = 'Invalid ESL Offset IDs',
  INVALID_ESL_USER = 'User should be an ESL Teacher',
  INVALID_FIRST_NAME = 'Invalid First Name',
  INVALID_LAST_NAME = 'Invalid Last Name',
  INVALID_LEAVE = "Leave doesn't exist",
  INVALID_LEAVE_ID = 'Invalid Leave ID',
  INVALID_LEAVE_TYPE = 'Invalid leave type',
  INVALID_MANAGER = "Manager doesn't exist",
  INVALID_MIDDLE_NAME = 'Invalid Middle Name',
  INVALID_NON_ESL_USER = 'User should not be an ESL Teacher',
  INVALID_NOTIFICATION = 'Invalid Notification',
  INVALID_OVERTIME = "Overtime doesn't exist",
  INVALID_POSITION = 'Invalid Position',
  INVALID_PROJECT = "Project doesn't exist",
  INVALID_PROJECT_LEADER = "Project Leader doesn't exist",
  INVALID_ROLE = 'Invalid Role',
  INVALID_SCHEDULE = "Schedule doesn't exist",
  INVALID_SCHEDULE_ID = 'Employee schedule does not exist',
  INVALID_SCHEDULE_NAME = 'Invalid Schedule Name',
  INVALID_START_TIME = 'Invalid Start Time',
  INVALID_TEAM_LEADER = 'Invalid Team Leader',
  INVALID_TIME_ENTRY = "Time entry doesn't exist",
  INVALID_USER = "User doesn't exist",
  INVALID_USER_ID = 'Invalid user ID',
  LEAVE_NOT_PENDING = 'Leave is not pending anymore',
  MISSING_APPROVED_MINUTES = 'Approved minutes is required',
  MISSING_PROJECTS = 'Project/s is required',
  MISMATCH_MANAGER = "Manager doesn't match",
  MISMATCH_PROJECT_LEADER = "Project Leader doesn't match",
  MISSING_LEAVE_DATES = 'Leave Date/s is required',
  NOT_HR_ADMIN = 'User is not an HR Admin',
  NOT_MANAGER_PROJECT_LEADER = 'User is not Manager or Project Leader',
  REQUIRED_TIME_IN_REMARKS = 'Remarks is required',
}